+++ tests/internal/test_deps.py
@@ -1,8 +1,9 @@
 import platform
 import sys
+from pathlib import Path
 from unittest import mock
 
-import pkg_resources
+import importlib_metadata as metadata
 import pytest
 
 from mopidy.internal import deps
@@ -86,25 +87,31 @@ class TestDeps:
             result = deps.gstreamer_info()
             assert "    none" in result["other"]
 
-    @mock.patch("pkg_resources.get_distribution")
+    @mock.patch("importlib_metadata.distribution")
     def test_pkg_info(self, get_distribution_mock):
-        dist_setuptools = mock.Mock()
-        dist_setuptools.project_name = "setuptools"
+        dist_setuptools = mock.MagicMock()
+        dist_setuptools.name = "setuptools"
         dist_setuptools.version = "0.6"
-        dist_setuptools.location = "/tmp/example/setuptools"
-        dist_setuptools.requires.return_value = []
+        dist_setuptools.locate_file = mock.MagicMock(
+            return_value=Path("/tmp/example/setuptools/main.py")
+        )
+        dist_setuptools.requires = []
 
         dist_pykka = mock.Mock()
-        dist_pykka.project_name = "Pykka"
+        dist_pykka.name = "Pykka"
         dist_pykka.version = "1.1"
-        dist_pykka.location = "/tmp/example/pykka"
-        dist_pykka.requires.return_value = [dist_setuptools]
+        dist_pykka.locate_file = mock.MagicMock(
+            return_value=Path("/tmp/example/pykka/main.py")
+        )
+        dist_pykka.requires = [f"{dist_setuptools.name}==0.6"]
 
         dist_mopidy = mock.Mock()
-        dist_mopidy.project_name = "Mopidy"
+        dist_mopidy.name = "Mopidy"
         dist_mopidy.version = "0.13"
-        dist_mopidy.location = "/tmp/example/mopidy"
-        dist_mopidy.requires.return_value = [dist_pykka]
+        dist_mopidy.locate_file = mock.MagicMock(
+            return_value=Path("/tmp/example/mopidy/no_name.py")
+        )
+        dist_mopidy.requires = [f"{dist_pykka.name}==1.1"]
 
         get_distribution_mock.side_effect = [
             dist_mopidy,
@@ -126,9 +133,9 @@ class TestDeps:
         assert "setuptools" == dep_info_setuptools["name"]
         assert "0.6" == dep_info_setuptools["version"]
 
-    @mock.patch("pkg_resources.get_distribution")
+    @mock.patch("importlib_metadata.distribution")
     def test_pkg_info_for_missing_dist(self, get_distribution_mock):
-        get_distribution_mock.side_effect = pkg_resources.DistributionNotFound
+        get_distribution_mock.side_effect = metadata.PackageNotFoundError("test")
 
         result = deps.pkg_info()
 
@@ -136,10 +143,10 @@ class TestDeps:
         assert "version" not in result
         assert "path" not in result
 
-    @mock.patch("pkg_resources.get_distribution")
+    @pytest.mark.skip("Version control missing in metadata")
+    @mock.patch("importlib_metadata.distribution")
     def test_pkg_info_for_wrong_dist_version(self, get_distribution_mock):
-        get_distribution_mock.side_effect = pkg_resources.VersionConflict
-
+        # get_distribution_mock.side_effect = metadata.VersionConflict
         result = deps.pkg_info()
 
         assert "Mopidy" == result["name"]